Factor out and reuse RetryProvider as default Miner (not MinerClient) provider
