+++ packages/package-utils/package.json
@@ -12,6 +12,7 @@
     "axios": "^0.28.0",
     "discord.js": "^12.5.3",
     "ethers": "^5.6.9",
+    "exponential-backoff": "^3.1.0",
     "jsonfile": "^6.1.0",
     "slack": "^11.0.2"
   }
