+++ packages/reputation-miner/bin/index.js
@@ -5,9 +5,9 @@ const { argv } = require("yargs")
   .option('minerAddress', {string:true})
   .option('providerAddress', {type: "array", default: []});
 const ethers = require("ethers");
-const backoff = require("exponential-backoff").backOff;
 
 const ReputationMinerClient = require("../ReputationMinerClient");
+const RetryProvider = require("../../package-utils/retryProvider").default;
 
 const { ConsoleAdapter, SlackAdapter, DiscordAdapter, TruffleLoader } = require("../../package-utils");
 
@@ -30,32 +30,6 @@ const {
   processingDelay,
   adapterLabel,
 } = argv;
-
-class RetryProvider extends ethers.providers.StaticJsonRpcProvider {
-  constructor(connectionInfo, adapterObject){
-    super(connectionInfo);
-    this.adapter = adapterObject;
-  }
-
-  static attemptCheck(err, attemptNumber){
-    console.log("Retrying RPC request #", attemptNumber);
-    if (attemptNumber === 5){
-      return false;
-    }
-    return true;
-  }
-
-  getNetwork(){
-    return backoff(() => super.getNetwork(), {retry: RetryProvider.attemptCheck});
-  }
-
-  // This should return a Promise (and may throw erros)
-  // method is the method name (e.g. getBalance) and params is an
-  // object with normalized values passed in, depending on the method
-  perform(method, params) {
-    return backoff(() => super.perform(method, params), {retry: RetryProvider.attemptCheck, startingDelay: 1000});
-  }
-}
 
 if ((!minerAddress && !privateKey) || !colonyNetworkAddress || !syncFrom) {
   console.log("❗️ You have to specify all of ( --minerAddress or --privateKey ) and --colonyNetworkAddress and --syncFrom on the command line!");
